refactor: split start event in a module
